ssx/single_sharded: improve argument and ret val processing, improve tests
- make argument processing consistent with ss::sharded<Service>::invoke_on,
i.e. make them value moved or copied even if lvalue was provided, and only
pass on a reference via std::reference_wrapper
- accept a pointer to method as a callable
- tidy up the tests
