+++ src/v/ssx/tests/CMakeLists.txt
@@ -32,7 +32,7 @@ rp_test(
   SOURCES
     abort_source_test.cc
     sharded_ptr_test.cc
-    single_sharded.cc
+    single_sharded_test.cc
   LIBRARIES v::seastar_testing_main
   ARGS "-- -c 2"
   LABELS ssx
